requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
greenlet>=3.0

# Object storage client
minio==7.2.13

# Database migrations
alembic==1.14.0

# Kafka
kafka-python==2.0.2

